Created qr uploader reusable component; updated shards import to use new component
